requirements.txt
# Web Framework
Flask==3.0.2
Flask-CORS==4.0.0
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0
aiohttp>=3.9.1
aiohttp-cors>=0.7.0
requests>=2.31.0

# Data Processing
numpy==1.24.3
pandas==2.0.3
textblob>=0.17.1

# AI/ML
openai==1.0.0
googletrans-py==4.0.0

# Utilities
cachetools==5.3.1
psutil==5.9.5

# Machine Learning
xgboost==2.0.3
scikit-learn==1.4.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.3.1

# Type Checking
mypy==1.8.0
types-requests==2.31.0.20240125
types-python-dateutil==2.8.19.20240106

# Configuration
PyYAML==6.0.1
